move asset health
